fix: prevent crash if the backend has been moved/deleted.
We just had the issue that the backend in vault had been moved
or unmounted. In this case, the vault read returns nil and our
resourceRead crashes with a nilpointer-exception, even if
terraform would destroy + recreate the resource.

This marks the resource as deleted if the plugin cannot find the
configuration anymore.