fix: prefer env-var VAULT_TOKEN over .vault-token if set
During tests of the terraform provider, we ran into permission denied
responses from vault. This occured, because before this change, for
example a token helper from a ~/.vault or a token in ~/.vault-token
took precedence over a VAULT_TOKEN env var. In normal vault operations,
the VAULT_TOKEN env var should win and it does now.